+++ src/Resources/Font.h
@@ -15,7 +15,7 @@ public:
     bool loadFromMemory(MEByte* bytes, std::size_t size);
     sf::Font* getResourcePointer();
 
-    unsigned int getLineSpacing(const unsigned int& size) const;
+    float getLineSpacing(const unsigned int& size) const;
 
     /**
      * Default font - fallback font
